Add random number generation and enhance drawing operations
The `rand` crate was added to the project for generating random numbers. This is used in our opcode logic to provide for random number generation in CHIP-8 operations. Also, a significant amount of code was added to handle the CHIP-8 DRAW opcode. It handles pixel flipping, sprite wrapping around the screen, and updating the CF register.