Add screen drawing functionality to main.rs
The changes in main.rs file primarily involve the addition of a new function `draw_screen` which handles the SDL rendering. Display buffer data, provided by the emulated Chip8 system, is used to render individual pixels. The updates to the chip are now included in the main event loop. The changes to the lib.rs file simply involve cleaning up an unused test module.