Refactor the emulator and add SDL2 support for desktop
The emulator code ('chip8_core/src/lib.rs') was significantly refactored with changes made to most of the functions. Added support for SDL2 in ther 'desktop' project, including the addition of a new file ('build.rs') which specifies the SDL2 system path, and key mapping functions to 'main.rs'. This commit enhances the emulation accuracy and improves the desktop version's interactivity.